requirements.txt
asgiref==3.4.1
certifi==2021.5.30
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.6
django-easy-softdelete==0.1.0
django-filter==2.4.0
django-multiselectfield==0.1.12
django-rest-auth==0.9.5
django-rest-swagger==2.1.2
django-reversion==3.0.9
django-revision==0.3.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
gitdb==4.0.7
GitPython==3.1.18
idna==2.6
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
openapi-codec==1.3.2
PyJWT==1.7.1
pytz==2021.1
requests==2.18.4
revision==0.1.1
simplejson==3.17.5
six==1.11.0
smmap==4.0.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.22
